Add files via upload
My project name is "IRRIGATION SYSTEM FOR SMART FARMING".  Main aim of my is to reduce the farmers work and time. 
Agriculture is the primary occupation in our country. India’s major income source is 
depending on agriculture therefore the development of agriculture is important. In today 
also most of the irrigation system are operated manually. The available traditional 
techniques are like drip irrigation, sprinkler irrigation etc. These techniques are need to 
be combined with IoT so that we can make use of water vary efficiently and also to control 
the Distribution System for Canal Irrigation System. IoT helps to access information and 
make major decision-making process by getting different values from sensors like soil 
moisture, water level sensors, water quality etc. This paper focuses primarily on reducing 
the wastage of water and minimizing the manual labor on field for irrigation so that you 
can saving time, cash and power of the farmer.And also Automate the Distribution System 
for canal irrigation system.